+++ CHANGES.md
@@ -2,12 +2,21 @@
 
 ## CHANGELOG ##
 
+* 5.2.0 (2023-03-04)
+  * Most asserts now accepts optional message string as PHPUnit's.
+  * Added `assertArrayHasKeyValue()`
+  * `assertArrayElement()` is now deprecated.
+  * Corrected signature of `assertRFC3339OrNull()`.
+  * `printArray()`'s indent block string is now exposed as optional parameter.
+
+
 * 5.1.0 (2022-11-28)
   * Fixed `assertIsInteger()` referencing wrong constant.
   * Corrected the way `printaArray()` handles conversion exceptions.
   * Corrected `assertIsObjectOrExistingClass()` not passing exception to `assertIsType()`
   * Added `Generator` tests.
   * Added PHPStan Github Action.
+
 
 * 5.0.0 (2022-11-18)
   * `Generator::getRandomBool()` probability now works opposite, and sets the probability of
@@ -17,16 +26,20 @@
   * `Generator::getRandomStringOrNull()` probability now works opposite and defines the chance of
     having the `string` value returned.
 
+
 * 4.0.0 (2022-11-04)
   * Dropped support for PHP 7.
   * The `Bridge` class is no longer part of the package. Use `marcin-orlowski/php-lockpick` instead.
 
+*
 * 3.4.0 (2022-09-26)
   * Internal `Validator` and `Type` are replaced by ones provided by `TypeAsserts` package.
   * Exception classes are replaced by ones provided by `TypeAsserts` package.
 
+
 * 3.3.0 (2022-09-20)
   * Added `getRandomLongitude()` and `getRandomLatitude()` generators.
+
 
 * 3.2.0 (2022-09-20)
   * Fixed invalid signature of `getRandomFloat()`.
@@ -35,9 +48,11 @@
   * Added default configs for `PHPStan`, `markdownlint` and `pre-commit` checkers.
   * Updated documentation.
 
+
 * 3.1.0 (2022-09-11)
   * Added `Generator` class to factory random values of various types.
   * Added `Bridge` class to manipulate hidden methods/properties/constants.
+
 
 * 3.0.0 (2022-09-11)
   * [BREAKING] Changed package structure and method access.
@@ -46,6 +61,7 @@
   * Extra asserts are no longer a `Trait` but regular class.
   * All asserts are now `static` methods.
   * Incorporated `TypeValidator` and related data type asserts into this package.
+
 
 * 2.0.0 (2022-01-17)
   * [BREAKING] Signature of `assertArraysHaveDifferences()` changed.
@@ -58,23 +74,29 @@
     * `assertArraysEquals()` (use `assertArrayEquals()` instead).
   * Corrected code style.
 
+
 * 1.3.1 (2022-01-15)
   * Enforced global namespace for all PHP core method calls.
   * `printArray()` now properly deals with objects not implementing `__toString()`.
 
+
 * 1.3.0 (2021-12-31)
   * Added `assertArrayContains()` and `assertArrayNotContains()` asserts.
 
+
 * 1.2.1 (2021-01-21)
   * Code cleanup.
+
 
 * 1.2.0 (2020-10-20)
   * Deprecated `assertArraysEquals()`. Use `assertArrayEquals()` instead.
   * Updated documentation.
 
+
 * 1.1.0
   * Make all `assertXXX()` methods `public`.
   * Updated documentation.
+
 
 * 1.0.3
   * Corrected package namespace.